backend/requirements.txt
Flask==3.0.3
pymongo==4.7.3
flask-pymongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
